+++ INVENTARIO/urls.py
@@ -1,7 +1,29 @@
 from django.urls import path
+
+from INVENTARIO.api_views import TiposTributosListAPIView, TributoByTipoListAPIView, TributoDetailsAPIView, productosListAPIView
 from . import views
 
 urlpatterns = [
+
+
+    # -------------------------------
+    # Endpoints API REST
+    # -------------------------------
+
+
+    #---------- URLS TRIBUTOS ----------#
+    path('api/tipo-tributos/', TiposTributosListAPIView.as_view(), name='tipo_tributos_api'),
+    path('api/tributos/tipo/<int:tipo_valor>/', TributoByTipoListAPIView.as_view(), name='tributos_api'),
+    path('api/tributo/<int:pk>/', TributoDetailsAPIView.as_view(), name='tributo_api'), 
+
+
+    #----------- PRODUCTOS -----------#
+    #URL PRODUCTOS
+    path('api/productos/', productosListAPIView.as_view(), name='tipo_dte_api'), 
+
+    # Endpoints API REST
+    # -------------------------------
+
     # Productos
     path('productos/', views.listar_productos, name='productos-lista'),
     path('productos/crear/', views.crear_producto, name='crear-producto'),
