+++ intracoe-frontend/.env
@@ -1,1 +1,2 @@
 VITE_URL_BASE=http://127.0.0.1:8000/api/api
+VITE_URL_BASE_INVENT=http://127.0.0.1:8000/inventario/api
