+++ intracoe/settings.py
@@ -30,6 +30,8 @@ ALLOWED_HOSTS=['*']
 
 CORS_ALLOWED_ORIGINS = [
     'http://localhost:5173',  # Asegúrate de que esta URL esté incluida
+    'http://localhost:5174',  # Asegúrate de que esta URL esté incluida
+    
 ]
 
 CORS_ALLOW_CREDENTIALS = True
@@ -53,6 +55,9 @@ CSRF_TRUSTED_ORIGINS = [
     'https://192.168.100.81',
     'http://localhost:5173',  # Puerto de tu frontend (React)
     'http://127.0.0.1:5173',  # Alternativa para localhost
+    'http://localhost:5174',  # Puerto de tu frontend (React)
+    'http://127.0.0.1:5174',  # Alternativa para localhost
+    
 ]
 
 ## CONFIGURACION DE CORREOS ##############################
